test: inherit a global tests/conftest.py via --confcutdir; drop fixture re-imports
Each suite's rootdir is its own dir, so a parent conftest was never discovered —
suites re-imported shared fixtures from tests/fixtures.py with a noqa. Rename
fixtures.py -> tests/conftest.py and add --confcutdir=.. to each suite's
pytest.toml so the parent conftest is inherited directly; delete the two
import-only suite conftests. The task-backend reset is now a single global
autouse at the parent (every suite mutates settings.TASKS).
