+++ tests/conftest.py
@@ -1,9 +1,26 @@
 import pytest
+from asgiref.local import Local
 from django.contrib.auth import get_user_model
 from django.core.exceptions import ImproperlyConfigured
 from django.db.utils import OperationalError, ProgrammingError
+from django.tasks import task_backends
 
 from django_absurd.queues import get_absurd_client
+
+
+@pytest.fixture(autouse=True)
+def _reset_task_backends():
+    """Blow away the task-backend cache before each test so a mutated ``settings.TASKS``
+    can't leak a stale backend into the next test's task resolution.
+
+    ``TaskBackendHandler`` caches the ``TASKS`` setting (a cached_property) and each
+    created backend (in ``_connections``); Django 6.0's test setting_changed receivers
+    do NOT reset it, so a stale backend would bind tasks to the wrong queues and make
+    task resolution non-deterministic across tests. Every suite mutates
+    ``settings.TASKS``, so this reset lives at the shared parent conftest.
+    """
+    task_backends._connections = Local(task_backends.thread_critical)
+    task_backends.__dict__.pop("settings", None)
 
 
 @pytest.fixture(autouse=True)
