--- tests/multidb/conftest.py
@@ -1,4 +1,0 @@
-from tests.fixtures import (  # noqa: F401
-    _enable_db,
-    _reset_absurd_queues,
-)
